requirements.txt
keybert
langchain
langchain-community
langchain-huggingface
langchain-openai
langchain-pinecone
numpy
openai
pillow
pinecone-client
pydub
reportlab
requests
scikit-learn
speechrecognition
streamlit
torch
transformers
whisper
yt-dlp